Data on 2,000 Spanish train trips

This data set is a random subsample of a much larger database of trips taken on the Spanish High Speed Train Service (Renfe AVE).

SPrail

Format

A data frame with 2,000 rows and 9 variables:

insert_date

Date and time when ticket was paid for.

origin

Origin City

destination

Destination City

start_date

Date and time for train departure.

end_date

Date and time for train arrival.

train_type

Train service name.

price

Price of ticket in Euros.

train_class

Class of ticket: tourist, business, etc.. Variable in Spanish.

fare

Type of ticket fare.

Source

https://www.kaggle.com/thegurus/spanish-high-speed-rail-system-ticket-pricing

Details

All dates and times are European Central Time.

The larger data set from which SPrail was sampled was compiled and released under GPL-2 public license by Pedro Muñoz and David Cañones.