Join two pibbles together

These are generic functions that dispatch to individual pibble methods. pibble structure from x will be maintained. pibble structure from y will be lost. See join for complete documentation.

# S3 method for tbl_pb
left_join(x, y, by = NULL, copy = FALSE, suffix = c(".x", ".y"), ...)

# S3 method for tbl_pb
inner_join(x, y, by = NULL, copy = FALSE, suffix = c(".x", ".y"), ...)

# S3 method for tbl_pb
right_join(x, y, by = NULL, copy = FALSE, suffix = c(".x", ".y"), ...)

# S3 method for tbl_pb
full_join(x, y, by = NULL, copy = FALSE, suffix = c(".x", ".y"), ...)

# S3 method for tbl_pb
semi_join(x, y, by = NULL, copy = FALSE, ...)

# S3 method for tbl_pb
nest_join(x, y, by = NULL, copy = FALSE, keep = FALSE, name = NULL, ...)

# S3 method for tbl_pb
anti_join(x, y, by = NULL, copy = FALSE, ...)

Arguments

x
A pair of data frames, data frame extensions (e.g. a tibble), or lazy data frames (e.g. from dbplyr or dtplyr). See Methods, below, for more details.

y
A pair of data frames, data frame extensions (e.g. a tibble), or lazy data frames (e.g. from dbplyr or dtplyr). See Methods, below, for more details.

by
A character vector of variables to join by.

If NULL, the default, *_join() will perform a natural join, using all variables in common across x and y. A message lists the variables so that you can check they're correct; suppress the message by supplying by explicitly.

To join by different variables on x and y, use a named vector. For example, by = c("a" = "b") will match x$a to y$b.

To join by multiple variables, use a vector with length > 1. For example, by = c("a", "b") will match x$a to y$a and x$b to y$b. Use a named vector to match different variables in x and y. For example, by = c("a" = "b", "c" = "d") will match x$a to y$b and x$c to y$d.

To perform a cross-join, generating all combinations of x and y, use by = character().

copy
If x and y are not from the same data source, and copy is TRUE, then y will be copied into the same src as x. This allows you to join tables across srcs, but it is a potentially expensive operation so you must opt into it.

suffix
If there are non-joined duplicate variables in x and y, these suffixes will be added to the output to disambiguate them. Should be a character vector of length 2.

...
Other parameters passed onto methods.

keep
Should the join keys from both x and y be preserved in the output? Only applies to nest_join(), left_join(), right_join(), and full_join().

name
The name of the list column nesting joins create. If NULL the name of y is used.