Functions to perform standard panel-data calculations

WARNING: FOR NOW, ALWAYS SPECIFY THE PANEL STRUCTURE EVERY TIME YOU CALL THIS FUNCTION. PDECLARE STATUS IS DROPPED BY MOST FUNCTIONS.

within_i(.var, .df = get(".", envir = parent.frame()),
  .fcn = function(x) mean(x, na.rm = TRUE), .i = NA, .t = NA,
  .d = NA, .uniqcheck = FALSE)

between_i(.var, .df = get(".", envir = parent.frame()),
  .fcn = function(x) mean(x, na.rm = TRUE), .i = NA, .t = NA,
  .d = NA, .uniqcheck = FALSE)

Arguments

.var
Vector to be transformed

.df
Data frame or tibble (usually the data frame or tibble that contains .var) which contains the panel structure variables either listed in .i and .t, or earlier declared with as_pdeclare(). If tlag is called inside of a dplyr verb, this can be omitted and the data will be picked up automatically.

.fcn
The function to be passed to dplyr::summarize(). x - .fcn(x) within .i is the within tranformation. .fcn(x) within .i minus .fcn overall is the between transformation. This will almost always be the default .fcn = function(x) mean(x,na.rm=TRUE).

.i
Character or character vector with the variable names that identify the individual cases. Note that setting any one of .i, .t, or .d will override all three already applied to the data, and will return data that is as_pdeclare()d with all three, unless .setpanel=FALSE.

.t
Character variable with the single variable name indicating the time. pmdplyr accepts two kinds of time variables: numeric variables where a fixed distance .d will take you from one observation to the next, or, if .d=0, any standard variable type with an order. Consider using the time_variable() function to create the necessary variable if your data uses a Date variable for time.

.d
Number indicating the gap in .t between one period and the next. For example, if .t indicates a single day but data is collected once a week, you might set .d=7. To ignore gap length and assume that "one period ago" is always the most recent prior observation in the data, set .d=0. By default, .d=1.

.uniqcheck
Logical parameter. Set to TRUE to always check whether .i and .t uniquely identify observations in the data. By default this is set to FALSE and the check is only performed once per session, and only if at least one of .i, .t, or .d is set.

Details

Thesese functions perform the standard between and within transformations on panel data.

Examples

data(SPrail)
# Calculate within- and between-route variation in price and add it to the data
SPrail <- SPrail %>%
  dplyr::mutate(
    within_route = within_i(price, .i = c("origin", "destination")),
    between_route = between_i(price, .i = c("origin", "destination"))
  )