Function to convert between panel data types

This function takes panel data objects declared using pmdplyr (pibble/tbl_pb), tsibble (tsibble/tbl_ts), plm (pdata.frame), and panelr (panel_data) and converts to one of the other three formats for use with functions in those packages.

panel_convert(data, to, ...)

Arguments

data
Data frame - a pibble, tsibble, pdata.frame, or panel_data object.

to
Character variable set to "pmdplyr", "pibble", "tbl_pb", "tsibble", "tbl_ts", "plm", "pdata.frame", "panelr" or "panel_data" indicating the type/package to be converted to.

...
Additional arguments to be sent to, respectively, as_pibble(), as_tsibble(), pdata.frame(), or panel_data().

Details

Any grouping will be lost. You must have the relevant package installed to convert to the type for that package. Conversions from pdata.frame will be improved if sjlabelled is also installed.

When using panel_convert, be aware of the requirements that each type has:

Feature/Requirement	|	pibble

|	tsibble		|	pdata.frame

|	panel_data

ID		.i

key			index[1]

id

Time		.t

index			index[2]

wave			Gap control

.d			regular

No	No		ID must exist

No	No	Yes	Yes

Time must exist	No	Yes	Yes	Yes[1]

Only one ID variable	No	No	Yes

Yes		Unique identification	No	Yes

No[2]	No[2]			Feature/Requirement

[1] pdata.frame does not require that time be provided, but if not provided will create it based on original ordering of the data. The pdata.frame option to set index equal to an integer for a balanced panel and have it figure out the rest by itself is not supported.

[2] pdata.frame and panel_data do not require that ID and time uniquely identify the observations on declaring the data, but functions in these packages may not work correctly without unique identification.

In addition to the above, be aware that the different packages have different requirements on which variable classes can be Time variables. pmdplyr::time_variable() can build an integer variable that will work in all packages.

You may run into some trouble if your data contains variables by the names panel_convert_id, panel_convert_time, pibble_d, or panel_convert_regular.

Examples

# Examples are set to not run in case you don't have the relevant target package installed.
if (interactive()) {
  data(Scorecard)

  S_pibble <- as_pibble(Scorecard, .i = unitid, .t = year)
  S_tsibble <- tsibble::as_tsibble(Scorecard, key = unitid, index = year)
  S_pdata.frame <- plm::pdata.frame(Scorecard, index=c("unitid", "year"))
  S_panel_data <- panelr::panel_data(Scorecard, id = unitid, wave = year)

  # Make everything a pibble
  # Note variable order may be different coming from panelr, which changes variable order
  S_pibble
  panel_convert(S_tsibble, to = "pibble")
  panel_convert(S_pdata.frame, to = "pibble")
  panel_convert(S_panel_data, to = "pibble")

  # Make everything a tsibble
  panel_convert(S_pibble, to = "tsibble")
  S_tsibble
  panel_convert(S_pdata.frame, to = "tsibble")
  panel_convert(S_panel_data, to = "tsibble")

  # Now for pdata.frame
  head(panel_convert(S_pibble, to = "plm"))
  head(panel_convert(S_tsibble, to = "plm"))
  head(S_pdata.frame)
  head(panel_convert(S_panel_data, to = "plm"))

  # And finally panel_data
  panel_convert(S_pibble, to = "panelr")
  panel_convert(S_tsibble, to = "panelr")
  panel_convert(S_pdata.frame, to = "panelr")
  S_panel_data

}