pibble methods

These are variants of existing functions that are designed to retain the pibble status of the object, as well as its .i, .t, and .d attributes.

# S3 method for tbl_pb
mutate(.data, ...)

# S3 method for tbl_pb
mutate_all(.tbl, .funs, ...)

# S3 method for tbl_pb
mutate_at(.tbl, .vars, .funs, ..., .cols = NULL)

# S3 method for tbl_pb
mutate_if(.tbl, .predicate, .funs, ...)

# S3 method for tbl_pb
distinct(.data, ..., .keep_all = FALSE)

# S3 method for tbl_pb
group_by(.data, ...)

# S3 method for tbl_pb
group_by_all(.tbl, .funs = list(), ..., .add = FALSE,
  .drop = dplyr::group_by_drop_default(.tbl))

# S3 method for tbl_pb
group_by_all(.tbl, .funs = list(), ..., .add = FALSE,
  .drop = dplyr::group_by_drop_default(.tbl))

# S3 method for tbl_pb
group_by_at(.tbl, .vars, .funs = list(), ...,
  .add = FALSE, .drop = dplyr::group_by_drop_default(.tbl))

# S3 method for tbl_pb
group_by_if(.tbl, .predicate, .funs = list(), ...,
  .add = FALSE, .drop = dplyr::group_by_drop_default(.tbl))

# S3 method for tbl_pb
ungroup(x, ...)

# S3 method for tbl_pb
bind_cols(.data, ...)

# S3 method for tbl_pb
intersect(x, y, ...)

# S3 method for tbl_pb
union(x, y, ...)

# S3 method for tbl_pb
union_all(x, y, ...)

# S3 method for tbl_pb
setdiff(x, y, ...)

# S3 method for tbl_pb
left_join(x, y, by = NULL, copy = FALSE,
  suffix = c(".x", ".y"), ...)

# S3 method for tbl_pb
inner_join(x, y, by = NULL, copy = FALSE,
  suffix = c(".x", ".y"), ...)

# S3 method for tbl_pb
right_join(x, y, by = NULL, copy = FALSE,
  suffix = c(".x", ".y"), ...)

# S3 method for tbl_pb
full_join(x, y, by = NULL, copy = FALSE,
  suffix = c(".x", ".y"), ...)

# S3 method for tbl_pb
semi_join(x, y, by = NULL, copy = FALSE, ...)

# S3 method for tbl_pb
nest_join(x, y, by = NULL, copy = FALSE,
  keep = FALSE, name = NULL, ...)

# S3 method for tbl_pb
anti_join(x, y, by = NULL, copy = FALSE, ...)

# S3 method for tbl_pb
select(.data, ...)

# S3 method for tbl_pb
select_all(.tbl, .funs = list(), ...)

# S3 method for tbl_pb
select_at(.tbl, .vars, .funs = list(), ...)

# S3 method for tbl_pb
select_if(.tbl, .predicate, .funs = list(), ...)

# S3 method for tbl_pb
rename(.data, ...)

# S3 method for tbl_pb
rename_all(.tbl, .funs = list(), ...)

# S3 method for tbl_pb
rename_at(.tbl, .vars, .funs = list(), ...)

# S3 method for tbl_pb
rename_if(.tbl, .predicate, .funs = list(), ...)

# S3 method for tbl_pb
summarize(.data, ...)

# S3 method for tbl_pb
summarize_all(.tbl, .funs, ...)

# S3 method for tbl_pb
summarize_at(.tbl, .vars, .funs, ...)

# S3 method for tbl_pb
summarize_if(.tbl, .predicate, .funs, ...)

# S3 method for tbl_pb
summarise(.data, ...)

# S3 method for tbl_pb
summarise_all(.tbl, .funs, ...)

# S3 method for tbl_pb
summarise_at(.tbl, .vars, .funs, ...)

# S3 method for tbl_pb
summarise_if(.tbl, .predicate, .funs, ...)

# S3 method for tbl_pb
transmute(.data, ...)

# S3 method for tbl_pb
transmute_all(.tbl, .funs, ...)

# S3 method for tbl_pb
transmute_at(.tbl, .vars, .funs, ..., .cols = NULL)

# S3 method for tbl_pb
transmute_if(.tbl, .predicate, .funs, ...)

Arguments

.data, x, .tbl
These functions take a tbl_pb (i.e. pibble) object as input

.add, by, .cols, copy, .drop, .funs, keep, .keep_all, name, .predicate, suffix, .vars, y, ...
Other parameters to be passed to the relevant functions

Details

Some functions that already preserve pibble status and so don't need special methods include:

dplyr::filter, dplyr::filter_all, dplyr::filter_at, dplyr::filter_if, dplyr::arrange, dplyr::arrange_all, dplyr::arrange_at, dplyr::arrange_if, dplyr::sample_frac, dplyr::slice, dplyr::sample_n, dplyr::top_n, dplyr::add_row, tibble:add_column

dplyr::bind_rows is currently not supported.

Any function that takes two data frames/tibbles as inputs will retain the panel structure of the first argument.

If a function is not on the above list or elsewhere in this help file, then you may need to re-as_pibble your object after using the function.