Join two pibbles together

These are generic functions that dispatch to individual pibble methods. See join for complete documentation.

# S3 method for tbl_pb
left_join(x, y, by = NULL, copy = FALSE,
  suffix = c(".x", ".y"), ...)

# S3 method for tbl_pb
inner_join(x, y, by = NULL, copy = FALSE,
  suffix = c(".x", ".y"), ...)

# S3 method for tbl_pb
right_join(x, y, by = NULL, copy = FALSE,
  suffix = c(".x", ".y"), ...)

# S3 method for tbl_pb
full_join(x, y, by = NULL, copy = FALSE,
  suffix = c(".x", ".y"), ...)

# S3 method for tbl_pb
semi_join(x, y, by = NULL, copy = FALSE, ...)

# S3 method for tbl_pb
nest_join(x, y, by = NULL, copy = FALSE,
  keep = FALSE, name = NULL, ...)

# S3 method for tbl_pb
anti_join(x, y, by = NULL, copy = FALSE, ...)

Arguments

x
tbls to join

y
tbls to join

by
a character vector of variables to join by. If NULL, the default, *_join() will do a natural join, using all variables with common names across the two tables. A message lists the variables so that you can check they're right (to suppress the message, simply explicitly list the variables that you want to join).

To join by different variables on x and y use a named vector. For example, by = c("a" = "b") will match x.a to y.b.

copy
If x and y are not from the same data source, and copy is TRUE, then y will be copied into the same src as x. This allows you to join tables across srcs, but it is a potentially expensive operation so you must opt into it.

suffix
If there are non-joined duplicate variables in x and y, these suffixes will be added to the output to disambiguate them. Should be a character vector of length 2.

...
other parameters passed onto methods, for instance, na_matches to control how NA values are matched. See join.tbl_df for more.

keep
If TRUE the by columns are kept in the nesting joins.

name
the name of the list column nesting joins create. If NULL the name of y is used.